Replaced with ECN code

+++ src/wifi/model/yans-error-rate-model.h
@@ -21,6 +21,8 @@
 #ifndef YANS_ERROR_RATE_MODEL_H
 #define YANS_ERROR_RATE_MODEL_H
 
+#include <stdint.h>
+#include "wifi-mode.h"
 #include "error-rate-model.h"
 #include "dsss-error-rate-model.h"
 
@@ -54,10 +56,6 @@ namespace ns3 {
 class YansErrorRateModel : public ErrorRateModel
 {
 public:
-  /**
-   * \brief Get the type ID.
-   * \return the object TypeId
-   */
   static TypeId GetTypeId (void);
 
   YansErrorRateModel ();
@@ -83,7 +81,7 @@ private:
    *
    * \return BER of BPSK at the given SNR
    */
-  double GetBpskBer (double snr, uint32_t signalSpread, uint64_t phyRate) const;
+  double GetBpskBer (double snr, uint32_t signalSpread, uint32_t phyRate) const;
   /**
    * Return BER of QAM-m with the given parameters.
    *
@@ -94,7 +92,7 @@ private:
    *
    * \return BER of BPSK at the given SNR
    */
-  double GetQamBer (double snr, unsigned int m, uint32_t signalSpread, uint64_t phyRate) const;
+  double GetQamBer (double snr, unsigned int m, uint32_t signalSpread, uint32_t phyRate) const;
   /**
    * Return k!
    *
@@ -145,7 +143,7 @@ private:
    * \return double
    */
   double GetFecBpskBer (double snr, double nbits,
-                        uint32_t signalSpread, uint64_t phyRate,
+                        uint32_t signalSpread, uint32_t phyRate,
                         uint32_t dFree, uint32_t adFree) const;
   /**
    * \param snr SNR ratio (not dB)
@@ -161,7 +159,7 @@ private:
    */
   double GetFecQamBer (double snr, uint32_t nbits,
                        uint32_t signalSpread,
-                       uint64_t phyRate,
+                       uint32_t phyRate,
                        uint32_t m, uint32_t dfree,
                        uint32_t adFree, uint32_t adFreePlusOne) const;
 };
